requirements.txt
discord-py-slash-command
discord.py
dnspython>=2.2
docker>=5.0
emoji
pylint>=2.13
pymongo>=4.1
requests
nltk